# internal/api/announcer/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Announcer public API",
    "version": "2.0",
    "contact": {
      "name": "E",
      "url": "https://blog.bullgare.com/2020/07/complete-list-of-swagger-options-to-protobuf-file/"
    }
  },
  "tags": [
    {
      "name": "AnnounceService"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v2/announce/add_ip": {
      "post": {
        "summary": "add Ip from ino net interface",
        "operationId": "AnnounceService_AddIP",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/announcerAddIpRequest"
            }
          }
        ],
        "tags": [
          "AnnounceService"
        ]
      }
    },
    "/v2/announce/remove_ip": {
      "post": {
        "summary": "remove Ip from net interface",
        "operationId": "AnnounceService_RemoveIP",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/announcerRemoveIpRequest"
            }
          }
        ],
        "tags": [
          "AnnounceService"
        ]
      }
    },
    "/v2/announce/state": {
      "get": {
        "summary": "get current state",
        "operationId": "AnnounceService_GetState",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/announcerGetStateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "AnnounceService"
        ]
      }
    }
  },
  "definitions": {
    "announcerAddIpRequest": {
      "type": "object",
      "properties": {
        "ip": {
          "type": "string",
          "title": "IP адрес"
        }
      },
      "title": "AddIpRequest запрос на добавление Ip в net-интерфейс"
    },
    "announcerGetStateResponse": {
      "type": "object",
      "properties": {
        "services": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "title": "GetStateResponse запрос на получение списка адресов из интефейса"
    },
    "announcerRemoveIpRequest": {
      "type": "object",
      "properties": {
        "ip": {
          "type": "string",
          "title": "IP адрес"
        }
      },
      "title": "RemoveIpRequest запрос на удаление Ip из net-интерфейса"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}